created types with mapping
